Apply previous review recommendations
